ExifToolHelper() - added terminate() overload
moved "not so well tested" functionality out to an ExifToolAlpha class, which will be an extension of ExifToolHelper

This code has not been tested yet.  The tests will be re-targeted to this multi layer class approach soon.
